Update to uber/h3 v3.5.0 and refactor test code
