+++ README.md
@@ -5,7 +5,7 @@
 [![Coverage Status](https://coveralls.io/repos/github/uber/h3-java/badge.svg?branch=master)](https://coveralls.io/github/uber/h3-java?branch=master)
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](LICENSE)
 [![Maven Central](https://maven-badges.herokuapp.com/maven-central/com.uber/h3/badge.svg)](https://maven-badges.herokuapp.com/maven-central/com.uber/h3)
-[![H3 Version](https://img.shields.io/badge/h3-v3.4.3-blue.svg)](https://github.com/uber/h3/releases/tag/v3.4.3)
+[![H3 Version](https://img.shields.io/badge/h3-v3.5.0-blue.svg)](https://github.com/uber/h3/releases/tag/v3.5.0)
 
 This library provides Java bindings for the [H3 Core Library](https://github.com/uber/h3). For API reference, please see the [H3 Documentation](https://uber.github.io/h3/).
 
@@ -17,14 +17,14 @@ Add it to your pom.xml:
 <dependency>
     <groupId>com.uber</groupId>
     <artifactId>h3</artifactId>
-    <version>3.4.1</version>
+    <version>3.5.0</version>
 </dependency>
 ```
 
 Or, using Gradle:
 
 ```gradle
-compile("com.uber:h3:3.4.1")
+compile("com.uber:h3:3.5.0")
 ```
 
 Encode a location into a hexagon address:
