+++ h3version.properties
@@ -1,1 +1,1 @@
-h3.git.reference=v3.4.3
+h3.git.reference=v3.5.0
